@@ -6,16 +6,16 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^0.1.3
-  path_provider: ^1.6.7
-  fluttertoast: ^7.1.1
-  flare_flutter: ^2.0.4
-  image_downloader: ^0.20.1
+  cupertino_icons: ^1.0.2
+  path_provider: ^2.0.1
+  fluttertoast: ^8.0.4
+  flare_flutter: ^3.0.0
+  image_downloader: ^0.30.0
 
 dev_dependencies:
   flutter_test:
@@ -36,4 +36,3 @@ flutter:
       fonts:
         - asset: assets/fonts/open-sans-extrabold.ttf
           weight: 700
-
