@@ -1,5 +1,4 @@
 //
-
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:fluttertoast/fluttertoast.dart';
@@ -20,7 +19,7 @@ class Utils{
     return AVATAR_BASE_URL + userId + "&size=" + "1024";
   }
 
-  Future<void> downloadImage(BuildContext context, String url, {AndroidDestinationType destination, bool whenError = false, String outputMimeType}) async {
+  Future<void> downloadImage(BuildContext context, String url, {AndroidDestinationType? destination, bool whenError = false, String? outputMimeType}) async {
     String fileName;
     String path;
     int size;
@@ -29,7 +28,7 @@ class Utils{
       String imageId;
 
       if (whenError) {
-        imageId = await ImageDownloader.downloadImage(url, outputMimeType: outputMimeType).catchError((error) {
+        imageId = (await ImageDownloader.downloadImage(url, outputMimeType: outputMimeType).catchError((error) {
           if (error is PlatformException) {
             var path = "";
             if (error.code == "404") {
@@ -44,35 +43,35 @@ class Utils{
           print("timeout");
           Fluttertoast.showToast(msg: "timeout");
           return;
-        });
+        }))!;
       } else {
         if (destination == null) {
-          imageId = await ImageDownloader.downloadImage(
+          imageId = (await ImageDownloader.downloadImage(
             url,
             outputMimeType: outputMimeType,
-          );
+          ))!;
         } else {
-          imageId = await ImageDownloader.downloadImage(
+          imageId = (await ImageDownloader.downloadImage(
             url,
             destination: destination,
             outputMimeType: outputMimeType,
-          );
+          ))!;
         }
       }
 
       if (imageId == null) {
         return;
       }
-      fileName = await ImageDownloader.findName(imageId);
-      path = await ImageDownloader.findPath(imageId);
-      size = await ImageDownloader.findByteSize(imageId);
-      mimeType = await ImageDownloader.findMimeType(imageId);
+      fileName = (await ImageDownloader.findName(imageId))!;
+      path = (await ImageDownloader.findPath(imageId))!;
+      size = (await ImageDownloader.findByteSize(imageId))!;
+      mimeType = (await ImageDownloader.findMimeType(imageId))!;
 
       if(path != null){
         Navigator.of(context).push(MaterialPageRoute(builder: (_) => AvatarPreviewScreen(filePath: path,))).then((value) => Navigator.pop(context));
       }
     } on PlatformException catch (error) {
-      Fluttertoast.showToast(msg: error.message);
+      Fluttertoast.showToast(msg: error.message!);
       print(error.message);
       return;
     }
